% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ggplot.lm}
\alias{ggplot.lm}
\title{Method for plotting linear models}
\usage{
ggplot.lm(data, mapping, vars, ...)
}
\arguments{
\item{data}{The linear model object from \code{lm}}

\item{mapping}{Regular mapping, see \code{ggplot} and \code{aes} for details.}

\item{vars}{A list of variable values used for prediction.}

\item{\ldots}{Additional arguments passed to \code{ggplot}}
}
\value{
A ggplot class object.
}
\description{
Simple method to plot a linear model using ggplot
along with 95% confidence intervals.
}
