% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Function to display descrptives statistics of a dataframe}
\usage{
numeric_summary(x, na.rm)
}
\arguments{
\item{x}{A vector of numeric values}

\item{na.rm}{A boolean value TRUE or FALSE}
}
\value{
a dataframe contains one row and 4 columns min, median, sd, max
}
\description{
THis function takes two arguments, a vector x and the logical na.rm
and return numeric summary of data
}
\examples{
# Example usage of numeric_summary
numeric_summary(airquality$Ozone, na.rm = TRUE)
}
\author{
Josue AFOUDA
}
