% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{Numbers and letters.}
\format{
A data frame with 10 rows and 2 variables
}
\source{
Created as a sample
}
\usage{
sampleData
}
\description{
A sample data set containing 10 numbers and letters with two variables:
}
\details{
\itemize{
  \item Num. A number.
  \item Letter. An upper case letter (A to J)
}
}
\keyword{datasets}
